+++ src/adapters/ui/SceneRouter.tsx
@@ -11,6 +11,7 @@ import { SceneTransition, type SceneTransitionRef } from './SceneTransition'
 import { UpdateNotification } from './UpdateNotification'
 import { LicenseToast } from './LicenseToast'
 import { TrialBadge } from './TrialBadge'
+import { WindowTitleBar } from './WindowTitleBar'
 
 type ActiveScene = 'free' | 'pomodoro' | 'fureai' | 'gallery'
 
@@ -60,6 +61,7 @@ export function SceneRouter(): JSX.Element {
       <UpdateNotification pomodoroActive={activeScene === 'pomodoro'} />
       <LicenseToast licenseMode={licenseMode} serverMessage={serverMessage} />
       <TrialBadge licenseMode={licenseMode} />
+      <WindowTitleBar />
     </>
   )
 }